update deploy script

+++ deploy_production.sh
@@ -28,8 +28,10 @@ yarn build
 
 echo "#### kill older server"
 PID=`ps ax | grep '[p]uma' | awk '{ print $1 }'`
-echo "older server is ${PID}"
-kill -9 ${PID}
+if [ ${PID} != "" ]; then
+  echo "older server is ${PID}"
+  kill -9 ${PID}
+fi
 
 echo "##### start server"
 RAILS_ENV=production bundle exec rails s -d
